requirements.txt
black
flake8
flask
flask-cors
flask-session
flask-marshmallow
marshmallow
PyMySQL==1.1.0
python-dotenv==1.0.0
redis
sentence-transformers==3.0.1
SQLAlchemy==2.0.30
tidb-vector==0.0.9



